APP file in this folder contains the main file>
The app file has the main code of the website. The website is can be used for saving data for customers students or any other. The data is stored in the main DB and displayed at the webpage further more then you can edit the entry after adding it in the DB first when it is displayed then you can click the edit button and then add or change information about the entry or the person.